+++ src/searchForm/autoComplete/overlay/AutoCompleteOverlay.tsx
@@ -2,21 +2,31 @@ import React from 'react'
 import Bullet, { BulletTypes } from 'bullet'
 import { AutoCompleteProps } from 'autoComplete'
 
-export interface AutoCompleteOverlayProps {
-  readonly name: string
-  readonly autocompleteComponent: React.ReactElement<AutoCompleteProps>
-  readonly className?: Classcat.Class
-  readonly onSelect?: (obj: AutocompleteOnChange) => void
+export type AutoCompleteOverlayProps = Omit<
+  AutoCompleteProps,
+  | 'autoFocus'
+  | 'inputAddon'
+  | 'embeddedInSearchForm'
+  // Adding searchForItems because AutoComplete needs to be passed from the
+  // renderAutocompleteComponent method
+  | 'searchForItems'
+> & {
+  readonly renderAutocompleteComponent: (
+    props: Omit<AutoCompleteProps, 'searchForItems'>,
+  ) => JSX.Element
 }
 
-export const AutoCompleteOverlay = (props: AutoCompleteOverlayProps) => {
+export const AutoCompleteOverlay = ({
+  renderAutocompleteComponent,
+  ...props
+}: AutoCompleteOverlayProps) => {
   const icon = (
     <div className="kirk-autoComplete-icon">
       <Bullet type={BulletTypes.SEARCH} />
     </div>
   )
 
-  return React.cloneElement(props.autocompleteComponent, {
+  return renderAutocompleteComponent({
     ...props,
     autoFocus: true,
     inputAddon: icon,